requirements.txt
flask>=2.2
requests>=2.28
gunicorn>=20.1
google-generativeai>=0.3.0
pytest>=7.0
